+++ main.go
@@ -559,7 +559,7 @@ func main() {
 
 	aC(Command{
 		name: "math",
-		re:   rl(`(?i)^!(math|quickmafs) (.*)$`),
+		re:   rl(`(?i)^` + prefi + `(math|quickmafs) (.*)$`),
 		callback: func(c *CommandEvent) {
 			exprString := c.Match[0][2]