+++ main.go
@@ -23,7 +23,7 @@ import (
 
 // Build Infos
 var (
-	Version = "3.5.0"
+	Version = "3.6.0"
 )
 
 // Flags