+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * answer regal with lager if nightbot writes it
 * scambot reply
-* the ability to leave multiple voicemails by joining the recipients with ` and `
+* the ability to leave multiple voicemails by joining the recipients with ` && `
 
 ### Changed