+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* show location on osm not gmaps
+
 ## [3.4.0] - 2019-12-12
 
 ### Added