+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * teamtrees command
 * misc unused commands
 
+
 ## [3.5.0] - 2019-12-18
 
 ### Added