+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * new command system
 * fill command
 
+### Changed
+
+* allow wetter as alias for weather
+
 ### Removed
 
 * teamtrees command